refactor: move `Layer` to new file (#96)
