+++ src/types/devices/host.ts
@@ -1,4 +1,4 @@
-import { Device, DeviceType, Layer } from "./device";
+import { Device, DeviceType } from "./device";
 import { ViewGraph } from "../graphs/viewgraph";
 import PcImage from "../../assets/pc.svg";
 import { Position } from "../common";
@@ -7,6 +7,7 @@ import { createDropdown, DeviceInfo, RightBar } from "../../graphics/right_bar";
 import { ProgramInfo } from "../../graphics/renderables/device_info";
 import { sendPacket } from "../packet";
 import { Ticker } from "pixi.js";
+import { Layer } from "./layer";
 
 const DEFAULT_ECHO_DELAY = 250; // ms
